refactor(storage): split `ReadObject` builder
